+++ resource/__init__.py
@@ -1,4 +1,5 @@
 from resource.agent import *
+from resource.algorithm import *
 from resource.connection import *
 from resource.dashboard import *
 from resource.data import *
@@ -16,6 +17,7 @@ __all__ = [
 db = [
     Agent_Catalog_Resource, Agent_Catalog_Selected_Resource,
     Agent_Instance_Resource, Agent_Instance_Selected_Resource,
+    Algorithm_Resource, Algorithm_Selected_Resource,
     Connection_Resource, Connection_Selected_Resource,
     Dashboard_Resource,
     Data_Resource, Data_Selected_Resource,