+++ about.py
@@ -1,4 +1,4 @@
 project = 'ASTRID'
 title = 'Context Broker Manager'
-version = '1.0.2'
+version = '1.0.3'
 description = 'Get and update collected data of the service chain with topology information.'