+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - Fix error to get exact match in query.
 - Improved execution of action as daemon.
 - Force execution of request even if the LCP operation failed.
+- Accept list of strings for eBPF-Program code.