+++ CHANGELOG.md
@@ -21,3 +21,5 @@
 - Added Algorithm management.
 - Fixed heartbeat period.
 - Stdout and stderr response format: plain, lines, and JSON.
+
+## 1.0.3